# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Redoubt API
  description: |
    Redoubt is a secure, invite-only communication platform API.

    All error responses conform to RFC 9457 Problem Details for HTTP APIs.
    All timestamps are in ISO 8601 UTC format.
    All IDs are UUID v4.
  version: 0.1.0
  contact:
    name: Redoubt
  license:
    name: MIT
servers:
  - url: https://api.example.com/api/v1
    description: Production server
  - url: http://localhost:8080/api/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    Timestamp:
      type: string
      format: date-time
      example: "2026-02-22T10:30:00Z"

    # Error schemas (RFC 9457)
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type
          example: "https://redoubt.app/problems/validation-error"
        title:
          type: string
          description: Short, human-readable summary
          example: "Validation Failed"
        status:
          type: integer
          description: HTTP status code
          example: 400
        detail:
          type: string
          description: Human-readable explanation
          example: "The request body contains invalid fields"
        instance:
          type: string
          description: URI reference for this occurrence
          example: "/api/v1/auth/register"
        request_id:
          type: string
          description: X-Request-ID for tracing
          example: "abc-123-def"
        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: "username"
        message:
          type: string
          example: "must be between 3 and 32 characters"

    # Auth schemas
...
    Channel:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        space_id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          example: "general"
        type:
          type: string
          enum: [text, voice]
        position:
          type: integer
          example: 0
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
    CreateChannelRequest:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          pattern: "^[a-zA-Z0-9 ]+$"
          example: "general"
        type:
          type: string
          enum: [text, voice]
        position:
          type: integer
          description: Position in channel list (auto-assigned if not specified)
    UpdateChannelRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
          pattern: "^[a-zA-Z0-9 ]+$"
    ReorderChannelsRequest:
      type: object
      required:
        - channel_ids
      properties:
        channel_ids:
          type: array
          items:
            $ref: '#/components/schemas/UUID'
          description: Channel IDs in desired order
    ChannelsListResponse:
      type: object
      properties:
        channels:
          type: array
          items:
            $ref: '#/components/schemas/Channel'

    # Message schemas
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
paths:
...
  /spaces/{id}/channels:
    get:
      tags: [Channels]
      summary: List channels
      description: List all channels in a space. Requires space membership.
      operationId: listChannels
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: List of channels
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelsListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Channels]
      summary: Create channel
      description: Create a new channel in a space. Requires admin role.
      operationId: createChannel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChannelRequest'
      responses:
        '201':
          description: Channel created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /spaces/{id}/channels/reorder:
    patch:
      tags: [Channels]
      summary: Reorder channels
      description: Reorder channels in a space. Requires admin role.
      operationId: reorderChannels
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderChannelsRequest'
      responses:
        '200':
          description: Channels reordered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelsListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /channels/{id}:
    get:
      tags: [Channels]
      summary: Get channel
      description: Get channel details. Requires space membership.
      operationId: getChannel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Channel details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Channels]
      summary: Update channel
      description: Update channel details. Requires admin role.
      operationId: updateChannel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChannelRequest'
      responses:
        '200':
          description: Channel updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Channels]
      summary: Delete channel
      description: Soft delete a channel. Requires admin role.
      operationId: deleteChannel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Channel deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Message endpoints (nested under channels)
...